==> 2_purpose.Rmd <==
---
title: "Anaylises for SONATA review"
subtitle: "Purpose field"
author: "Alicia Valdés"
date: "`r format(Sys.time(), '%d %B %Y')`"
output:
  pdf_document: default
  html_notebook: default
---

This is a script to perform analyses for the SONATA review paper.

# Load libraries

```{r}
library(here)
library(readxl)
library(tidyverse)
```

# Read data

```{r}
sonata_data <-read_excel(here("data", "raw","database_sonata_v2.2.xlsx"),
                         sheet = "SONATAT1.1")
sonata_data
```

# Plots for Purpose

```{r}
sonata_data %>%
  # Set Purpose to factor and Year to integer
  mutate(Year = as.integer(Year), Purpose_AV = as.factor(Purpose)) %>%
  count(Year, Purpose) %>%
  ggplot(aes(x = Year, y = n, color = Purpose)) + geom_line(linewidth = 1) +
  theme_bw()
```

```{r}
sonata_data  %>%
  # Set Purpose to factor and Year to integer
  mutate(Year = as.integer(Year), Purpose = as.factor(Purpose)) %>%
  group_by(Year, Purpose) %>%
  summarise(n = n(), .groups = "drop") %>%
  group_by(Year) %>%
  mutate(total = sum(n), proportion = n / total) %>%
  ggplot(aes(x = Year, y = proportion, color = Purpose, size = total)) +
  geom_line(linewidth = 1) +
  scale_y_continuous(labels = scales::percent_format()) +
  labs(x = "Year", y = "Proportion of Articles", color = "Purpose") +
  theme_bw()
```

```{r}
sonata_data %>%
  # Set Purpose to factor and Year to integer
  mutate(Year = as.integer(Year), Purpose = as.factor(Purpose)) %>%
  group_by(Year, Purpose) %>%
  summarise(n = n(), .groups = "drop") %>%
  ggplot(aes(x = Year, y = n, fill = Purpose)) +
  geom_area(alpha = 0.8, color = "black") +
  labs(x = "Year", y = "Number of Articles", ill = "Purpose") +
  theme_bw()
```

```{r}
sonata_data %>%
  # Set Purpose to factor and Year to integer
  mutate(Year = as.integer(Year), Purpose = as.factor(Purpose)) %>%
  group_by(Year, Purpose) %>%
  summarise(n = n(), .groups = "drop") %>%
  group_by(Year) %>%
  mutate(proportion = n / sum(n)) %>%
  ggplot(aes(x = Year, y = proportion, fill = Purpose)) +
  geom_area(position = "fill", alpha = 0.8, color = "black") +
  scale_y_continuous(labels = scales::percent_format()) +
  labs(
    x = "Year",
    y = "Proportion of Articles",
    fill = "Purpose"
  ) +
  theme_bw()
```

```{r}
sonata_data %>%
  # Set Purpose to factor and Year to integer
  mutate(Year = as.integer(Year), Purpose = as.factor(Purpose)) %>%
  ggplot(aes(x = Year, fill = Purpose)) +
  geom_density(alpha = 0.8, position = "stack", bw = 0.5) +
  labs(x = "Year", y = "Density", fill = "Purpose") +
  theme_bw()
```

# Session info

```{r}
sessionInfo()
```
